+++ src/main/java/com/blackrook/rookscript/tools/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * This package contains some user tools.
+ */
+package com.blackrook.rookscript.tools;