+++ docs/CHANGELOG.md
@@ -3,6 +3,13 @@ RookScript (C) Black Rook Software
 by Matt Tropiano et al. (see AUTHORS.txt)
 
 
+Changed in 1.1.0
+----------------
+
+- `Added` ScriptExecutor main class.
+- `Changed` Exception thrown in ScriptValue.createForType(...) if conversion absolutely fails.
+
+
 Changed in 1.0.0
 ----------------